splitToPeriods fix

+++ pinfin/model/src/commonMain/kotlin/hnau/pinfin/model/utils/HueExt.kt
@@ -1,9 +1,9 @@
 package hnau.pinfin.model.utils
 
 import hnau.pinfin.data.Hue
-import hnau.common.app.model.theme.Hue as ModelHue
+import hnau.common.app.model.utils.Hue as ModelHue
 
 val Hue.model: ModelHue
     get() = ModelHue(
-        degrees = degrees.toDouble(),
+        degrees = degrees,
     )
